e_os file moved to the "internal" include tree
Storing it with the other includes in place of the project root directory looks more appropriate.

All the files that includes it were update accordingly:
 include "e_os.h" -> include "internal/e_os.h"
