+++ apps/apps.h
@@ -10,7 +10,7 @@
 #ifndef HEADER_APPS_H
 # define HEADER_APPS_H
 
-# include "e_os.h" /* struct timeval for DTLS */
+# include "internal/e_os.h" /* struct timeval for DTLS */
 # include "internal/nelem.h"
 # include <assert.h>
