+++ crypto/bio/bio_lcl.h
@@ -7,7 +7,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include "e_os.h"
+#include "internal/e_os.h"
 #include "internal/sockets.h"
 #include "internal/refcount.h"
