+++ crypto/cryptlib.c
@@ -8,7 +8,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include "e_os.h"
+#include "internal/e_os.h"
 #include "internal/cryptlib_int.h"
 #include <openssl/safestack.h>
