+++ crypto/des/cfb_enc.c
@@ -7,7 +7,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include "e_os.h"
+#include "internal/e_os.h"
 #include "des_locl.h"
 #include <assert.h>
