+++ crypto/dh/dh_kdf.c
@@ -7,7 +7,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include "e_os.h"
+#include "internal/e_os.h"
 
 #ifndef OPENSSL_NO_CMS
 #include <string.h>
