+++ crypto/dllmain.c
@@ -7,7 +7,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include "e_os.h"
+#include "internal/e_os.h"
 #include "internal/cryptlib_int.h"
 
 #if defined(_WIN32) || defined(__CYGWIN__)
