+++ crypto/dso/dso_win32.c
@@ -7,7 +7,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include "e_os.h"
+#include "internal/e_os.h"
 #include "dso_locl.h"
 
 #if defined(DSO_WIN32)
