+++ crypto/engine/eng_devcrypto.c
@@ -7,7 +7,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include "e_os.h"
+#include "internal/e_os.h"
 #include <string.h>
 #include <sys/types.h>
 #include <sys/stat.h>
