+++ crypto/engine/eng_lib.c
@@ -7,7 +7,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include "e_os.h"
+#include "internal/e_os.h"
 #include "eng_int.h"
 #include <openssl/rand.h>
 #include "internal/refcount.h"
