+++ crypto/mem_sec.c
@@ -15,7 +15,7 @@
  * For details on that implementation, see below (look for uppercase
  * "SECURE HEAP IMPLEMENTATION").
  */
-#include "e_os.h"
+#include "internal/e_os.h"
 #include <openssl/crypto.h>
 
 #include <string.h>
