+++ crypto/o_str.c
@@ -7,7 +7,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include "e_os.h"
+#include "internal/e_os.h"
 #include <limits.h>
 #include <openssl/crypto.h>
 #include "internal/cryptlib.h"
