+++ crypto/ocsp/ocsp_ht.c
@@ -7,7 +7,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include "e_os.h"
+#include "internal/e_os.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include "internal/ctype.h"
