+++ ssl/ssl_init.c
@@ -7,7 +7,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include "e_os.h"
+#include "internal/e_os.h"
 
 #include "internal/err.h"
 #include <openssl/crypto.h>
